Replace Gutenberg Block Processor with the Tree Processor implementation
