+++ WordPress/Classes/ViewRelated/Gutenberg/Processors/GutenbergBlockProcessor.swift
@@ -17,36 +17,15 @@ public class GutenbergBlockProcessor: Processor {
     ///
     public typealias Replacer = (GutenbergBlock) -> String?
 
-    // MARK: - Basic Info
-
     let name: String
-
-    // MARK: - Regex
 
     private enum CaptureGroups: Int {
         case all = 0
         case name
         case attributes
-        case content
 
-        static let allValues: [CaptureGroups] = [.all, .name, .attributes, .content]
+        static let allValues: [CaptureGroups] = [.all, .name, .attributes]
     }
-
-    /// Regular expression to detect attributes
-    /// Capture groups:
-    ///
-    /// 1. The block id
-    /// 2. The block attributes
-    /// 3. Block content
-    ///
-    private lazy var gutenbergBlockRegexProcessor: RegexProcessor = { [weak self]() in
-        let pattern = "\\<!--[ ]?(\(name))([\\s\\S]*?)-->([\\s\\S]*?)<!-- \\/\(name) -->"
-        let regex = try! NSRegularExpression(pattern: pattern, options: .caseInsensitive)
-
-        return RegexProcessor(regex: regex) { (match: NSTextCheckingResult, text: String) -> String? in
-            return self?.process(match: match, text: text)
-        }
-    }()
 
     // MARK: - Parsing & processing properties
     private let replacer: Replacer
@@ -58,36 +37,137 @@ public class GutenbergBlockProcessor: Processor {
         self.replacer = replacer
     }
 
+    /// Regular expression to detect attributes of the opening tag of a block
+    /// Capture groups:
+    ///
+    /// 1. The block id
+    /// 2. The block attributes
+    ///
+    var openTagRegex: NSRegularExpression {
+        let pattern = "\\<!--[ ]?(\(name))([\\s\\S]*?)-->"
+        return try! NSRegularExpression(pattern: pattern, options: .caseInsensitive)
+    }
+
+    /// Regular expression to detect the closing tag of a block
+    ///
+    var closingTagRegex: NSRegularExpression {
+        let pattern = "\\<!-- \\/\(name) -->"
+        return try! NSRegularExpression(pattern: pattern, options: .caseInsensitive)
+    }
+
     // MARK: - Processing
 
+    /// Processes the block and for any needed replacements from a given opening tag match.
+    ///     - Parameters:
+    ///         - text: The string that the following parameter is found in.
+    ///     - Returns: The resulting string after the necessary replacements have occured
+    ///
     public func process(_ text: String) -> String {
-        return gutenbergBlockRegexProcessor.process(text)
+        let matches = openTagRegex.matches(in: text, options: [], range: text.utf16NSRange(from: text.startIndex ..< text.endIndex))
+        var replacements = [(NSRange, String)]()
+
+        var lastReplacementBound = 0
+        for match in matches {
+            if match.range.lowerBound >= lastReplacementBound, let replacement = process(match, in: text) {
+                replacements.append(replacement)
+                lastReplacementBound = replacement.0.upperBound
+            }
+        }
+        let resultText = replace(replacements, in: text)
+        return resultText
+    }
+
+    /// Replaces the
+    ///     - Parameters:
+    ///         - replacements: An array of tuples representing first a range of text that needs to be replaced then the string to replace
+    ///         - text: The string to perform the replacements on
+    ///
+    func replace(_ replacements: [(NSRange, String)], in text: String) -> String {
+        let mutableString = NSMutableString(string: text)
+        var offset = 0
+        for (range, replacement) in replacements {
+            let lengthBefore = mutableString.length
+            let offsetRange = NSRange(location: range.location + offset, length: range.length)
+            mutableString.replaceCharacters(in: offsetRange, with: replacement)
+            let lengthAfter = mutableString.length
+            offset += (lengthAfter - lengthBefore)
+        }
+        return mutableString as String
     }
 }
-
 // MARK: - Regex Match Processing Logic
 
 private extension GutenbergBlockProcessor {
-    /// Processes an Gutenberg block  regex match.
+    /// Processes the block and for any needed replacements from a given opening tag match.
+    ///     - Parameters:
+    ///         - match: The match reperesenting an opening block tag
+    ///         - text: The string that the following parameter is found in.
+    ///     - Returns: Any necessary replacements within the provided string
     ///
-    func process(match: NSTextCheckingResult, text: String) -> String? {
+    private func process(_ match: NSTextCheckingResult, in text: String) -> (NSRange, String)? {
 
-        guard match.numberOfRanges == CaptureGroups.allValues.count else {
+        var result: (NSRange, String)? = nil
+        if let closingRange = locateClosingTag(forMatch: match, in: text) {
+            let attributes = readAttributes(from: match, in: text)
+            let content = readContent(from: match, withClosingRange: closingRange, in: text)
+            let parsedContent = process(content) // Recurrsively parse nested blocks and process those seperatly
+            let block = GutenbergBlock(name: name, attributes: attributes, content: parsedContent)
+
+            if let replacement = replacer(block) {
+                let length = closingRange.upperBound - match.range.lowerBound
+                let range = NSRange(location: match.range.lowerBound, length: length)
+                result = (range, replacement)
+            }
+        }
+
+        return result
+    }
+
+    /// Determines the location of the closing block tag for the matching open tag
+    ///     - Parameters:
+    ///         - openTag: The match reperesenting an opening block tag
+    ///         - text: The string that the following parameter is found in.
+    ///     - Returns: The Range of the closing tag for the block
+    ///
+    func locateClosingTag(forMatch openTag: NSTextCheckingResult, in text: String) -> NSRange? {
+        guard let index = text.indexFromLocation(openTag.range.upperBound) else {
             return nil
         }
 
-        let attributes = readAttributes(from: match, in: text)
-        let content = readContent(from: match, in: text)
-        let block = GutenbergBlock(name: name, attributes: attributes, content: content)
+        let matches = closingTagRegex.matches(in: text, options: [], range: text.utf16NSRange(from: index ..< text.endIndex))
 
-        return replacer(block)
+        for match in matches {
+            let content = readContent(from: openTag, withClosingRange: match.range, in: text)
+
+            if tagsAreBalanced(in: content) {
+                return match.range
+            }
+        }
+
+        return nil
     }
 
-    // MARK: - Regex Match Processing Logic
-
-    /// Obtains the attributes from a block match.
+    /// Determines if there are an equal number of opening and closing block tags in the provided text.
+    ///     - Parameters:
+    ///         - text: The string to test assumes that a block with an even number represents a valid block sequence.
+    ///     - Returns: A boolean where true represents an equal number of opening and closing block tags of the desired type
     ///
-    private func readAttributes(from match: NSTextCheckingResult, in text: String) -> [String: Any] {
+    func tagsAreBalanced(in text: String) -> Bool {
+
+        let range = text.utf16NSRange(from: text.startIndex ..< text.endIndex)
+        let openTags = openTagRegex.matches(in: text, options: [], range: range)
+        let closingTags = closingTagRegex.matches(in: text, options: [], range: range)
+
+        return openTags.count == closingTags.count
+    }
+
+    /// Obtains the block attributes from a regex match.
+    ///     - Parameters:
+    ///         - match: The `NSTextCheckingResult` from a successful regex detection of an opening block tag
+    ///         - text: The string that the following parameter is found in.
+    ///     - Returns: A JSON dictionary of the block attributes
+    ///
+    func readAttributes(from match: NSTextCheckingResult, in text: String) -> [String: Any] {
         guard let attributesText = match.captureGroup(in: CaptureGroups.attributes.rawValue, text: text),
             let data = attributesText.data(using: .utf8 ),
             let json = try? JSONSerialization.jsonObject(with: data, options: .allowFragments),
@@ -98,13 +178,22 @@ private extension GutenbergBlockProcessor {
         return jsonDictionary
     }
 
-    /// Obtains the block content from a block match.
+    /// Obtains the block content from a regex match and range.
+    ///     - Parameters:
+    ///         - match: The `NSTextCheckingResult` from a successful regex detection of an opening block tag
+    ///         - closingRange: The `NSRange` of the closing block tag
+    ///         - text: The string that the following parameters are found in.
+    ///     - Returns: The content between the opening and closing tags of a block
     ///
-    private func readContent(from match: NSTextCheckingResult, in text: String) -> String {
-        guard let content = match.captureGroup(in: CaptureGroups.content.rawValue, text: text) else {
+    func readContent(from match: NSTextCheckingResult, withClosingRange closingRange: NSRange, in text: String) -> String {
+        guard let index = text.indexFromLocation(match.range.upperBound) else {
             return ""
         }
 
-        return content
+        guard let closingBound = text.indexFromLocation(closingRange.lowerBound) else {
+            return ""
+        }
+
+        return String(text[index..<closingBound])
     }
 }
